add warp_into function to GeoImage and remove the old WarpToImage function in gip_gdal header and cpp
